requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
plotly>=5.18.0
